Extract `//cuttlefish/host/commands/metrics:send`
Bug: b/447695201
