+++ base/cvd/cuttlefish/host/commands/metrics/utils.cc
@@ -17,6 +17,7 @@
 
 #include <net/if.h>
 #include <netinet/in.h>
+#include <stdint.h>
 #include <string.h>
 #include <sys/ioctl.h>
 #include <sys/time.h>
@@ -25,17 +26,9 @@
 #include <chrono>
 #include <cstring>
 #include <ctime>
-#include <iostream>
+#include <string>
 
 #include <android-base/logging.h>
-#include <android-base/strings.h>
-#include <gflags/gflags.h>
-
-#include "cuttlefish/host/libs/metrics/metrics_defs.h"
-#include "cuttlefish/host/libs/web/http_client/curl_global_init.h"
-#include "cuttlefish/host/libs/web/http_client/curl_http_client.h"
-#include "cuttlefish/host/libs/web/http_client/http_client.h"
-#include "cuttlefish/host/libs/web/http_client/http_string.h"
 
 namespace cuttlefish::metrics {
 
@@ -129,52 +122,6 @@ uint64_t GetEpochTimeMs() {
   uint64_t milliseconds_since_epoch =
       std::chrono::duration_cast<std::chrono::milliseconds>(now).count();
   return milliseconds_since_epoch;
-}
-
-std::string ClearcutServerUrl(ClearcutServer server) {
-  switch (server) {
-    case ClearcutServer::kLocal:
-      return "http://localhost:27910/log";
-
-    case ClearcutServer::kStaging:
-      return "https://play.googleapis.com:443/staging/log";
-
-    case ClearcutServer::kProd:
-      return "https://play.googleapis.com:443/log";
-
-    default:
-      LOG(FATAL) << "Invalid host configuration";
-      return "";
-  }
-}
-
-MetricsExitCodes PostRequest(const std::string& output, ClearcutServer server) {
-  CurlGlobalInit curl_global_init;
-  std::unique_ptr<HttpClient> http_client = CurlHttpClient();
-  if (!http_client) {
-    return MetricsExitCodes::kMetricsError;
-  }
-  return PostRequest(*http_client, output, server);
-}
-
-MetricsExitCodes PostRequest(HttpClient& http_client, const std::string& output,
-                             ClearcutServer server) {
-  std::string clearcut_url = ClearcutServerUrl(server);
-
-  Result<HttpResponse<std::string>> http_res =
-      HttpPostToString(http_client, clearcut_url, output);
-  if (!http_res.ok()) {
-    LOG(ERROR) << "HTTP command failed: " << http_res.error().FormatForEnv();
-    return MetricsExitCodes::kMetricsError;
-  }
-
-  if (!http_res->HttpSuccess()) {
-    LOG(ERROR) << "Metrics message failed: [" << http_res->data << "]";
-    LOG(ERROR) << "http error code: " << http_res->http_code;
-    return MetricsExitCodes::kMetricsError;
-  }
-  LOG(INFO) << "Metrics posted to ClearCut";
-  return MetricsExitCodes::kSuccess;
 }
 
 }  // namespace cuttlefish::metrics
